+++ test/test_steady_state.jl
@@ -1,10 +1,4 @@
-# activate the test environment if needed
-using Pkg
-if ! ("BenchmarkTools" ∈ keys(Pkg.project().dependencies))
-    using TestEnv; TestEnv.activate()
-end
-
-using Test, BenchmarkTools, StaticArrays, LinearAlgebra, KiteUtils, Plots
+using StaticArrays, LinearAlgebra, KiteUtils
 using KiteModels, KitePodModels
 
 if ! @isdefined kcu
@@ -21,8 +15,9 @@ KiteModels.set_depower_steering(kps, kps.set.depower_offset/100.0, 0.0)
 height = sin(deg2rad(kps.set.elevation)) * kps.set.l_tether
 kps.v_wind .= kps.v_wind_gnd * calc_wind_factor(kps, height)
 kps.stiffness_factor = 0.04
-y0, yd0 = KiteModels.find_steady_state(kps, true)
+
+@time KiteModels.find_steady_state(kps, true)
 
 println("\nlift, drag    [N]  : $(KiteModels.lift_drag(kps))")
-println("\nSpring forces:")
-spring_forces(kps)
+# println("\nSpring forces:")
+# spring_forces(kps)
